@@ -14,18 +14,20 @@
   "dependencies": {
     "@apollo/client": "^3.6.2",
     "@elastic/elasticsearch": "^7.17.0",
-    "@emotion/cache": "^11.7.1",
-    "@emotion/react": "^11.9.0",
-    "@emotion/server": "^11.4.0",
-    "@emotion/styled": "^11.8.1",
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "^11.11.3",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.8",
-    "@mui/base": "^5.0.0-alpha.80",
-    "@mui/icons-material": "^5.6.2",
-    "@mui/lab": "^5.0.0-alpha.81",
-    "@mui/material": "^5.7.0",
+    "@mui/base": "^5.0.0-beta.33",
+    "@mui/icons-material": "^5.15.9",
+    "@mui/lab": "^5.0.0-alpha.165",
+    "@mui/material": "^5.15.9",
+    "@mui/material-nextjs": "^5.15.9",
+    "@mui/system": "^5.15.11",
     "@react-spring/web": "^9.7.3",
     "@redis/client": "^1.2.0",
-    "@sentry/nextjs": "^7.64.0",
+    "@sentry/nextjs": "^7.109.0",
     "@use-gesture/react": "^10.3.0",
     "autosuggest-highlight": "^3.3.4",
     "classnames": "^2.3.1",
@@ -39,15 +41,14 @@
     "image-size": "^1.0.1",
     "isomorphic-dompurify": "^0.18.0",
     "mongodb": "^5.5.0",
-    "next": "^12.2.3",
-    "next-transpile-modules": "^9.0.0",
+    "next": "^14.1.4",
     "node-fetch": "^3.2.4",
     "node-vibrant": "^3.2.1-alpha.1",
     "plaiceholder": "^2.3.0",
-    "react": "18.1.0",
+    "react": "^18.2.0",
     "react-device-detect": "^2.2.2",
-    "react-dom": "18.1.0",
-    "sharp": "^0.30.4",
+    "react-dom": "^18.2.0",
+    "sharp": "^0.33.3",
     "smoothscroll-polyfill": "^0.4.4",
     "swr": "^2.2.4",
     "url-file-size": "^1.0.5-1",
@@ -97,11 +98,6 @@
     "production": [
       "chrome 49",
       "opera 36"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
     ]
   }
 }
