@@ -1,8 +1,9 @@
-import { useRouter } from "next/router";
-import { ChangeEvent, useCallback, useEffect, useState } from "react";
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { ChangeEvent, useCallback, useEffect, useState, useMemo } from "react";
 
-import { omit } from "../../../../../helpers/omit";
-import { Nullable } from "../../../../../helpers/typings/utility-types";
+import { omit } from "@/helpers/omit";
+import { Nullable } from "@/helpers/typings/utility-types";
+
 import { fill } from "./fill";
 import { getFilterStringFromList } from "./getFilterStringFromList";
 import { FilterList } from "./type";
@@ -18,32 +19,31 @@ export type useFilterHookResult = {
 };
 
 export const useFilter = (label: string): useFilterHookResult => {
-  const { query, push: routerPush } = useRouter();
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+  const currentFilter = searchParams?.get(label);
   const [filterList, setFilterList] = useState<Nullable<FilterList>>(
-    fill(query[label]),
+    fill(currentFilter),
+  );
+  const params = useMemo(
+    () => new URLSearchParams(searchParams ?? {}),
+    [searchParams],
   );
 
   useEffect(() => {
-    setFilterList(fill(query[label]));
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [query[label]]);
+    setFilterList(fill(currentFilter));
+  }, [currentFilter]);
 
-  useEffect(
-    () => {
-      const queryFilter = getFilterStringFromList(filterList);
-      routerPush(
-        {
-          query: queryFilter
-            ? { ...omit(query, ["page"]), [label]: queryFilter }
-            : omit(query, ["page", label]),
-        },
-        undefined,
-        { shallow: true, scroll: false },
-      );
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [filterList, label],
-  );
+  useEffect(() => {
+    const queryFilter = getFilterStringFromList(filterList);
+
+    if (queryFilter) params.set(label, queryFilter);
+    else params.delete(label);
+    params.delete("page");
+
+    router.push(`${pathname}?${params.toString()}`, { scroll: false });
+  }, [filterList, label, params, pathname, router]);
 
   const handleOnChange = useCallback<HandleOnChange>(
     ({ target: { name } }, checked) => {
