@@ -1,4 +1,4 @@
-import { useRouter } from "next/router";
+import { useSearchParams } from "next/navigation";
 import useSWR from "swr";
 
 import {
@@ -11,28 +11,23 @@ import { Nullable } from "@/helpers/typings/utility-types";
 type SWRKey = FetcherData & SearchParams;
 
 export const useQuerySearch = (ssrData: SearchResponseFrontend) => {
-  const { text, categories, departments, page } = useRouter()
-    .query as SearchParams;
-  const textSearch = text?.trim();
+  const searchParams = useSearchParams();
+  const params = new URLSearchParams(searchParams ?? {});
+  const query: SearchParams = Object.fromEntries(params);
+
+  params.delete("text");
+
+  const url = `${process.env.NEXT_PUBLIC_API_ES_URL}?${params.toString()}`;
+  const text = query.text?.trim();
 
   const { data, isValidating, error } = useSWR<
     SearchResponseFrontend,
     string,
     Nullable<SWRKey>
-  >(
-    {
-      url: process.env.NEXT_PUBLIC_API_ES_URL,
-      text: textSearch,
-      departments,
-      categories,
-      page,
-    },
-    fetcherData,
-    {
-      keepPreviousData: true,
-      fallbackData: ssrData,
-    },
-  );
+  >({ url, text }, fetcherData, {
+    keepPreviousData: true,
+    fallbackData: ssrData,
+  });
 
   if (error !== undefined) {
     exceptionLog(error);
