@@ -1,6 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import classNames from "classnames";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import { Children, FC, cloneElement, ReactElement } from "react";
 
 import { Link, LinkProps } from "../Link/Link";
@@ -18,12 +18,13 @@ export const ActiveLink: FC<ActiveLinkProps> = ({
   disableHref,
   ...props
 }) => {
-  const { asPath } = useRouter();
+  const pathname = usePathname();
+
   const child = Children.only(children);
 
   const href =
     typeof props.href === "string" ? props.href : props.href.href ?? "";
-  const isActive = `${asPath}/`.indexOf(href) !== -1;
+  const isActive = `${pathname}/`.indexOf(href) !== -1;
   const linkClassName = classNames(className, { [activeClassName]: isActive });
 
   return disableHref ? (
