@@ -1,8 +1,10 @@
+"use client";
+
 /* eslint-disable @next/next/no-img-element */
 /* eslint-disable react/jsx-props-no-spreading */
 import { Box, Fade, Skeleton } from "@mui/material";
 import classNames from "classnames";
-import NextImage, { ImageProps as NextImageProps } from "next/future/image";
+import NextImage, { ImageProps as NextImageProps } from "next/image";
 import { FC, useCallback, useState, useRef, useEffect } from "react";
 
 import { Nullable } from "../../helpers/typings/utility-types";
@@ -39,6 +41,7 @@ export const Image: FC<ImageProps> = ({
   width,
   height,
   src,
+  style,
   ...props
 }) => {
   const [isLoaded, setLoaded] = useState(false);
@@ -86,10 +89,11 @@ export const Image: FC<ImageProps> = ({
         alt={alt}
         aria-hidden={!alt}
         src={src}
+        style={style}
         className={classNames(classes.image, className)}
         width={width}
         height={height}
-        onLoadingComplete={handleOnLoad}
+        onLoad={handleOnLoad}
         placeholder={blurDataURL ? "blur" : "empty"}
         blurDataURL={blurDataURL ?? undefined}
         {...props}
@@ -105,6 +109,7 @@ export const Image: FC<ImageProps> = ({
                 width,
                 height,
                 blurDataURL,
+                objectFit: style?.objectFit,
               })}`}
               className={classes.blur}
             />
