@@ -1,16 +1,23 @@
+/* eslint-disable no-nested-ternary */
+import { CSSProperties } from "@/helpers/typings/utility-types";
+
 type GetImageBlurSvg = {
   width: number;
   height: number;
   blurDataURL: string;
+  objectFit?: CSSProperties["objectFit"];
 };
-export const getImageBlurSvg = ({
-  width,
-  height,
-  blurDataURL,
-}: GetImageBlurSvg) => {
-  const std = "20";
-  const feComponentTransfer = blurDataURL.startsWith("data:image/jpeg")
-    ? `%3CfeComponentTransfer%3E%3CfeFuncA type='discrete' tableValues='1 1'/%3E%3C/feComponentTransfer%3E%`
-    : "";
-  return `%3Csvg xmlns='http%3A//www.w3.org/2000/svg' viewBox='0 0 ${width} ${height}'%3E%3Cfilter id='b' color-interpolation-filters='sRGB'%3E%3CfeGaussianBlur stdDeviation='${std}'/%3E${feComponentTransfer}%3C/filter%3E%3Cimage filter='url(%23b)' x='0' y='0' height='100%25' width='100%25' href='${blurDataURL}'/%3E%3C/svg%3E`;
-};
+export function getImageBlurSvg(param: GetImageBlurSvg) {
+  const { width, height, blurDataURL, objectFit } = param;
+  const std = 40;
+  const viewBox = width && height ? `viewBox='0 0 ${width} ${height}'` : "";
+  const preserveAspectRatio = viewBox
+    ? "none"
+    : objectFit === "contain"
+    ? "xMidYMid"
+    : objectFit === "cover"
+    ? "xMidYMid slice"
+    : "none";
+
+  return `%3Csvg xmlns='http://www.w3.org/2000/svg' ${viewBox}%3E%3Cfilter id='b' color-interpolation-filters='sRGB'%3E%3CfeGaussianBlur stdDeviation='${std}'/%3E%3CfeColorMatrix values='1 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 100 -1' result='s'/%3E%3CfeFlood x='0' y='0' width='100%25' height='100%25'/%3E%3CfeComposite operator='out' in='s'/%3E%3CfeComposite in2='SourceGraphic'/%3E%3CfeGaussianBlur stdDeviation='${std}'/%3E%3C/filter%3E%3Cimage width='100%25' height='100%25' x='0' y='0' preserveAspectRatio='${preserveAspectRatio}' style='filter: url(%23b);' href='${blurDataURL}'/%3E%3C/svg%3E`;
+}
