@@ -1,8 +1,10 @@
+"use client";
+
 import dynamic from "next/dynamic";
 
 export const DynamicOffer = dynamic(
   () => import("./Offer").then((res) => res.Offer),
   {
-    ssr: true,
+    ssr: false,
   },
 );
