@@ -1,4 +1,4 @@
-import { Maybe } from "../../../../../helpers/typings/utility-types";
+import { MaybeOrNullable } from "@/helpers/typings/utility-types";
 
-export const fill = (page: Maybe<string | string[]>): number =>
+export const fill = (page: MaybeOrNullable<string | string[]>): number =>
   typeof page === "string" ? parseInt(page, 10) : 1;
