@@ -1,10 +1,11 @@
+"use client";
+
 /* eslint-disable react/jsx-props-no-spreading */
 import {
   Pagination as MUIPagination,
   PaginationProps as MUIPaginationProps,
   PaginationItem,
 } from "@mui/material";
-import { useRouter } from "next/router";
 import { FC } from "react";
 
 import { Link } from "../Link/Link";
@@ -24,30 +25,23 @@ export const Pagination: FC<PaginationProps> = ({
   firstPageLink,
   page,
   ...props
-}) => {
-  const { query } = useRouter();
-
-  const currentPage =
-    page ?? ((typeof query.page === "string" && parseInt(query.page, 10)) || 1);
-
-  return (
-    <MUIPagination
-      count={count}
-      page={currentPage}
-      siblingCount={currentPage >= 5 ? 2 : 1}
-      shape="rounded"
-      {...props}
-      renderItem={(item) => (
-        <PaginationItem
-          component={Link}
-          href={
-            item.page === 1
-              ? `${firstPageLink ?? uri}`
-              : `${uri}/page/${item.page}`
-          }
-          {...item}
-        />
-      )}
-    />
-  );
-};
+}) => (
+  <MUIPagination
+    count={count}
+    page={page}
+    siblingCount={page ?? 1 >= 5 ? 2 : 1}
+    shape="rounded"
+    {...props}
+    renderItem={(item) => (
+      <PaginationItem
+        component={Link}
+        href={
+          item.page === 1
+            ? `${firstPageLink ?? uri}`
+            : `${uri}/page/${item.page}`
+        }
+        {...item}
+      />
+    )}
+  />
+);
