@@ -1,9 +1,7 @@
 import { QueryOptions, gql } from "@apollo/client";
-import useSWR, { Fetcher, SWRConfiguration } from "swr";
 
 import { client } from "@/lib/apollo/client";
 import { TransformBlocks } from "@/core/backend/transformBlocks/utils/type";
-import { FetcherGQLData, fetcherGQLData } from "@/helpers/fetcherGQLData";
 import { Nullable } from "@/helpers/typings/utility-types";
 import { columnsBlockGQL } from "@/components/blocks/Columns/utils/columnsGQL";
 import { embedBlockGQL } from "@/components/blocks/Embed/utils/embedGQL";
@@ -22,11 +20,10 @@ import { spacerBlockGQL } from "@/components/blocks/Spacer/utils/spacerGQL";
 import { tableBlockGQL } from "@/components/blocks/Table/utils/tableGQL";
 import { verseBlockGQL } from "@/components/blocks/Verse/utils/verseGQL";
 import { videoBlockGQL } from "@/components/blocks/Video/utils/videoGQL";
-import { REVALIDATE } from "@/constants";
 
 import { PostFieldsGQL, postFieldsGQL } from "../../getPosts/gql/postListGQL";
 
-type GetPostQueryVariables = {
+export type GetPostQueryVariables = {
   id: string | number;
   type: "DATABASE_ID" | "ID" | "SLUG" | "URI";
   isPreview?: boolean;
@@ -36,7 +33,7 @@ type PostSareComponentsFieldsGQL = {
   link: string;
 };
 
-type GetPostQuery = {
+export type GetPostQuery = {
   post: Nullable<
     PostSareComponentsFieldsGQL & PostFieldsGQL & { blocks: TransformBlocks[] }
   >;
@@ -50,7 +47,7 @@ const postSareComponentsFieldsGQL = {
   `,
 };
 
-const getPostDocument = gql`
+export const getPostDocument = gql`
   query GetPostDocument($id: ID!, $type: PostIdType, $isPreview: Boolean) {
     post(id: $id, idType: $type, asPreview: $isPreview) {
       ...postSareComponentsFieldsGQL
@@ -104,17 +101,3 @@ export const clientGetPostQuery = (
   const options = { query: getPostDocument, ...baseOptions };
   return client.query<GetPostQuery, GetPostQueryVariables>(options);
 };
-
-export const useGetPostQuery = (
-  variables: GetPostQueryVariables,
-  config?: SWRConfiguration<
-    GetPostQuery,
-    Error,
-    Fetcher<GetPostQuery, FetcherGQLData<GetPostQueryVariables>>
-  >,
-) =>
-  useSWR<GetPostQuery, Error, FetcherGQLData<GetPostQueryVariables>>(
-    { document: getPostDocument, variables },
-    fetcherGQLData,
-    { refreshInterval: REVALIDATE.PREVIEW_POLL_INTERVAL, ...config },
-  );
