@@ -3,7 +3,6 @@ import { FC } from "react";
 
 import { Nullable } from "@/helpers/typings/utility-types";
 import { Article, ArticleProps } from "@/components/Article/Article";
-import { SEO, SEOProp } from "@/components/SEO/SEO";
 import {
   Pagination,
   PaginationProps,
@@ -28,20 +27,20 @@ type RoutePageProps = {
   }>;
   pagination?: Nullable<PaginationProps>;
   pageNumber?: number;
-} & Pick<SEOProp, "domenTitle">;
+};
 
 export const RoutePage: FC<RoutePageProps> = ({
   page,
   childrenPage,
   pagination,
   pageNumber,
-  domenTitle,
 }) =>
   childrenPage === null ? (
     <Article
       title={page.title}
       blocks={page.blocks}
       href={page.href}
+      imageUrl={page.imageUrl}
       isPreview={page.isPreview}
     />
   ) : (
@@ -50,12 +49,6 @@ export const RoutePage: FC<RoutePageProps> = ({
       component="section"
       maxWidth="md"
       disableGutters>
-      {pageNumber && (
-        <SEO
-          domenTitle={domenTitle}
-          title={`${page.title} — Cтраница ${pageNumber}`}
-        />
-      )}
       <Box sx={sxHeaderBox}>
         <Typography align="center" variant="h1" gutterBottom={!!pageNumber}>
           {page.title}
