@@ -4,14 +4,16 @@
 
 import * as Sentry from "@sentry/nextjs";
 
+const enabled =
+  process.env.NEXT_PUBLIC_SENTRY_ENABLED === "true"
+    ? true
+    : process.env.NEXT_PUBLIC_SENTRY_ENABLED !== "false";
+
 Sentry.init({
-  dsn: "https://0453608349524a9b8f1f1056543c5bab@o397370.ingest.sentry.io/5898663",
+  dsn: "https://0453608349524a9b8f1f1056543c5bab@o397370.ingest.us.sentry.io/5898663",
 
-  // Adjust this value in production, or use tracesSampler for greater control
-  tracesSampleRate: 0.5,
+  tracesSampleRate: 0.3,
 
-  // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,
-
-  enabled: process.env.NODE_ENV === "production",
+  enabled,
 });
